refactor(server): extract battle-snapshots row + converters into persistence/battle-snapshots.ts (part of #1559) by claude
Moves the BattleSnapshotRow interface, normalizeBattleSnapshotStatus,
normalizeBattleSnapshotPlayerIds, and toBattleSnapshotRecord out of
persistence.ts into a dedicated module.

Also extracts the three shared column helpers they depend on
(normalizePlayerId, formatTimestamp, parseJsonColumn) into
persistence/column-helpers.ts, then re-imports them into persistence.ts
so the 180+ existing call sites keep working unchanged.

The MySqlPersistence class methods (saveBattleSnapshotStart,
saveBattleSnapshotResolution, settleInterruptedBattleSnapshot,
listBattleSnapshotsForPlayer) stay on the class for now — splitting a
class across files is a separate refactor.

No runtime behavior change. Refs #1559.
