+++ apps/server/src/persistence/index.ts
@@ -2,3 +2,5 @@ export * from "./mysql-tables";
 export * from "./types";
 export * from "./env-readers";
 export * from "./defaults";
+export * from "./column-helpers";
+export * from "./battle-snapshots";
